Split Winamp code
