+++ winamp/Makefile
@@ -27,12 +27,14 @@ TARGET_EXT_LIBS += $(LIBS_TARGET_EXT_LIBS)
 
 export CFLAGS LDFLAGS
 
+#SRC_SRCS = $(wildcard *.c)
+SRC_SRCS = in_vgmstream.c in_streamfile.c in_config.c
 
 ###############################################################################
 ### targets
 
 in_vgmstream: libvgmstream.a $(TARGET_EXT_LIBS) resource.o
-	$(CC) -shared -static-libgcc $(CFLAGS) "-DVERSION=\"`../version.sh`\"" in_vgmstream.c resource.o $(LDFLAGS) -o $(OUTPUT_WINAMP)
+	$(CC) -shared -static-libgcc $(CFLAGS) "-DVERSION=\"`../version.sh`\"" $(SRC_SRCS) resource.o $(LDFLAGS) -o $(OUTPUT_WINAMP)
 	$(STRIP) $(OUTPUT_WINAMP)
 
 resource.o: resource.rc resource.h
