+++ winamp/in_vgmstream.vcproj
@@ -176,6 +176,10 @@
 			UniqueIdentifier="{93995380-89BD-4b04-88EB-625FBE52EBFB}"
 			>
 			<File
+				RelativePath=".\in_vgmstream.h"
+				>
+			</File>
+			<File
 				RelativePath=".\resource.h"
 				>
 			</File>
@@ -195,6 +199,14 @@
 			Filter="cpp;c;cc;cxx;def;odl;idl;hpj;bat;asm;asmx"
 			UniqueIdentifier="{4FC737F1-C7A5-4376-A066-2A32D752A2FF}"
 			>
+			<File
+				RelativePath=".\in_config.c"
+				>
+			</File>
+			<File
+				RelativePath=".\in_streamfile.c"
+				>
+			</File>
 			<File
 				RelativePath=".\in_vgmstream.c"
 				>
