+++ winamp/in_vgmstream.vcxproj.filters
@@ -15,6 +15,9 @@
     </Filter>
   </ItemGroup>
   <ItemGroup>
+    <ClInclude Include="in_vgmstream.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
     <ClInclude Include="resource.h">
       <Filter>Header Files</Filter>
     </ClInclude>
@@ -25,6 +28,12 @@
     </ResourceCompile>
   </ItemGroup>
   <ItemGroup>
+    <ClCompile Include="in_config.c">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="in_streamfile.c">
+      <Filter>Source Files</Filter>
+    </ClCompile>
     <ClCompile Include="in_vgmstream.c">
       <Filter>Source Files</Filter>
     </ClCompile>
